Move cwt() and icwt() to object methods, and calculate scales from given time axis
Instead of making the user independently calculate 'scales' and
'sampf' and keep track of it outside of pyCWT, now a time axis is
passed to cwt() to be used to automatically calculate a reasonable set
of scales.

Instead of passing the MotherWavelet object to a package function
cwt(), cwt() becomes a method of the MotherWavelet class.  Similarily,
icwt() is moved to a method of the WaveletResult() class (formerly
named Wavelet).  Most of the initialization parameters are moved from
the object constructor to the cwt() method, as now the scales are
calculated "as needed" when a cwt() is to be done.

The Morlet cone of influence calculation is rewritten, based on
Torrence & Compo 1998 Fortran code.